Data collecting Challenge
This code is developed to demonstrate web-scraping and data analysis skills using Python and libraries such as Beautiful Soup and Splinter. The aim of this project is to collect, organize, and analyze information about Mars, specifically its news articles and weather data.

The first deliverable focuses on scraping titles and preview texts from Mars news articles. This is achieved by identifying relevant HTML elements, their class and id attributes, and parsing the information using Beautiful Soup.

The second deliverable involves scraping and analyzing Mars weather data, which exists in a table. The analysis reveals that the third month has the coldest minimum temperature on Mars, while the eighth month is the warmest. The atmospheric pressure is, on average, lowest in the sixth month and highest in the ninth. The distance from peak to peak is roughly 1425-750, or 675 days. This is equivalent to a Mars year, which is confirmed through internet search to be about 687 earth days.